feat: update example struct
