+++ src/core/parsed_commands.ts
@@ -1,19 +1,13 @@
 import { debuglog } from 'node:util';
-import { assert } from 'node:console';
-
-import { pick, omit } from 'lodash';
 import { Injectable, Container, Inject, ScopeEnum } from '@artus/core';
-import { Middlewares } from '@artus/pipeline';
-
-import { CommandMeta, CommandConfig, OptionMeta, OptionConfig, MiddlewareConfig, MiddlewareMeta } from '../types';
-import { parseArgvToArgs, parseArgvWithPositional, parseCommand, ParsedCommandStruct, Positional } from './parser';
-import { Command, EmptyCommand } from './command';
+import { parseArgvToArgs, parseArgvWithPositional } from './parser';
+import { Command } from './command';
 import { MetadataEnum } from '../constant';
 import { BinInfo } from './bin_info';
-import { isInheritFrom, formatToArray } from '../utils';
 import { errors, ArtusCliError } from '../errors';
+import { ParsedCommand } from './parsed_command';
+import { ParsedCommandTree } from './parsed_command_tree';
 
-const OPTION_SYMBOL = Symbol('ParsedCommand#Option');
 const TREE_SYMBOL = Symbol('ParsedCommand#Tree');
 const debug = debuglog('artus-cli#ParsedCommands');
 
@@ -36,282 +30,6 @@ export interface MatchResult {
   args: Record<string, any>;
 }
 
-export interface ParsedCommandOption {
-  location?: string;
-  commandConfig: CommandConfig;
-  parsedCommandInfo: ParsedCommandStruct;
-  optionConfig?: {
-    flagOptions: OptionConfig;
-    argumentOptions: OptionConfig;
-  };
-}
-
-/** Wrapper of command */
-export class ParsedCommand implements ParsedCommandStruct {
-  /** cmds.join(' ') */
-  uid: string;
-  /** the last element of cmds, 'bin dev' is 'dev', 'bin module test [baseDir]' is 'test' */
-  cmd: string;
-  /** convert command to array, like [ 'bin', 'dev' ] */
-  cmds: string[];
-  /** user defined in options but remove bin name */
-  command: string;
-  alias: string[];
-  enable: boolean;
-  demanded: Positional[];
-  optional: Positional[];
-  description: string;
-  examples: string[];
-  globalOptions?: OptionConfig;
-  flagOptions: OptionConfig;
-  argumentOptions: OptionConfig;
-  /** Command class location */
-  location?: string;
-
-  /** child commands */
-  childs: ParsedCommand[];
-  /** parent command */
-  parent: ParsedCommand | null;
-  /** inherit command */
-  inherit: ParsedCommand | null;
-
-  commandConfig: CommandConfig;
-  commandMiddlewares: Middlewares;
-  executionMiddlewares: Middlewares;
-
-  constructor(public clz: typeof Command, option: ParsedCommandOption) {
-    const { location, commandConfig, parsedCommandInfo, optionConfig } = option;
-    this.location = location;
-
-    // read from parsed_command
-    this.uid = parsedCommandInfo.uid;
-    this.command = parsedCommandInfo.command;
-    this.cmd = parsedCommandInfo.cmd;
-    this.cmds = parsedCommandInfo.cmds;
-    this.demanded = parsedCommandInfo.demanded;
-    this.optional = parsedCommandInfo.optional;
-
-    // read from option config
-    this.flagOptions = optionConfig?.flagOptions || {};
-    this.argumentOptions = optionConfig?.argumentOptions || {};
-    this.childs = [];
-    this.parent = null;
-    this.inherit = null;
-
-    // read from command config
-    this.commandConfig = commandConfig;
-    this.description = commandConfig.description || '';
-    this.examples = formatToArray(commandConfig.examples);
-    this.enable = typeof commandConfig.enable === 'boolean' ? commandConfig.enable : true;
-    this.alias = formatToArray(commandConfig.alias);
-
-    // middleware config
-    this.commandMiddlewares = [];
-    this.executionMiddlewares = [];
-  }
-
-  get options(): OptionConfig {
-    if (!this[OPTION_SYMBOL]) {
-      this[OPTION_SYMBOL] = { ...this.globalOptions, ...this.flagOptions };
-    }
-    return this[OPTION_SYMBOL];
-  }
-
-  get isRoot() {
-    return !this.parent;
-  }
-
-  get isRunable() {
-    return this.clz !== EmptyCommand && this.enable;
-  }
-
-  get depth() {
-    return this.cmds.length;
-  }
-
-  addMiddlewares(type: 'command' | 'execution', config: MiddlewareConfig) {
-    const { middleware, mergeType } = config;
-    const middlewares = type === 'command' ? this.commandMiddlewares : this.executionMiddlewares;
-    const extraMiddlewares = Array.isArray(middleware) ? middleware : [ middleware ];
-    // mergeType default is after
-    if (!mergeType || mergeType === 'after') {
-      middlewares.push(...extraMiddlewares);
-    } else {
-      middlewares.unshift(...extraMiddlewares);
-    }
-  }
-
-  updateGlobalOptions(opt: OptionConfig) {
-    this.globalOptions = { ...this.globalOptions, ...opt };
-    this[OPTION_SYMBOL] = null;
-  }
-}
-
-/** Parsed Command Tree */
-export class ParsedCommandTree {
-  /** root of command tree */
-  root: ParsedCommand;
-
-  /** command list, the key is command string used to match argv */
-  commands: Map<string, ParsedCommand>;
-
-  /** cache the instance of parsedCommand */
-  parsedCommandMap: Map<typeof Command, ParsedCommand>;
-
-  constructor(
-    private binName: string,
-    private commandList: (typeof Command)[],
-  ) {
-    this.build();
-  }
-
-  /** convert Command class to ParsedCommand instance */
-  private initParsedCommand(clz: typeof Command) {
-    const metadata = Reflect.getOwnMetadata(MetadataEnum.COMMAND, clz) as CommandMeta;
-    if (!metadata) return null;
-
-    const commandMeta = metadata;
-    const inheritClass = Object.getPrototypeOf(clz);
-    const inheritCommand = this.initParsedCommand(inheritClass);
-
-    let commandConfig = { ...commandMeta.config };
-
-    // mege command config with inherit command
-    if (inheritCommand && commandMeta.inheritMetadata !== false) {
-      const inheritCommandConfig = inheritCommand.commandConfig;
-      commandConfig = Object.assign({}, {
-        alias: inheritCommandConfig.alias,
-        command: inheritCommandConfig.command,
-        description: inheritCommandConfig.description,
-        parent: inheritCommandConfig.parent,
-      } satisfies CommandConfig, commandConfig);
-    }
-
-    // default command is main command
-    commandConfig.command = commandConfig.command || '$0';
-
-    // init parent
-    let command = commandConfig.command!;
-    if (commandConfig.parent) {
-      const parentParsedCommand = this.initParsedCommand(commandConfig.parent);
-      assert(parentParsedCommand, `parent ${commandConfig.parent?.name} is not a valid Command`);
-      command = parentParsedCommand!.cmds.concat(command).join(' ');
-    }
-
-    // avoid creating parsedCommand again.
-    if (this.parsedCommandMap.has(clz)) {
-      return this.parsedCommandMap.get(clz);
-    }
-
-    // parse command usage
-    const parsedCommandInfo = parseCommand(command, this.binName);
-
-    // split options with argument key and merge option info with inherit command
-    const optionMeta: OptionMeta | undefined = Reflect.getOwnMetadata(MetadataEnum.OPTION, clz);
-    const argumentsKey = parsedCommandInfo.demanded.concat(parsedCommandInfo.optional).map(pos => pos.cmd);
-    let flagOptions: OptionConfig = omit(optionMeta?.config || {}, argumentsKey);
-    let argumentOptions: OptionConfig = pick(optionMeta?.config || {}, argumentsKey);
-    if (inheritCommand && optionMeta?.inheritMetadata !== false) {
-      flagOptions = Object.assign({}, inheritCommand.flagOptions, flagOptions);
-      argumentOptions = Object.assign({}, inheritCommand.argumentOptions, argumentOptions);
-    }
-
-    const parsedCommand = new ParsedCommand(clz, {
-      location: commandMeta.location,
-      commandConfig,
-      parsedCommandInfo,
-      optionConfig: { flagOptions, argumentOptions },
-    });
-
-    if (inheritCommand) parsedCommand.inherit = inheritCommand;
-    if (this.commands.has(parsedCommandInfo.uid)) {
-      const existsParsedCommand = this.commands.get(parsedCommandInfo.uid)!;
-
-      // override only allow in class inheritance or options.override=true
-      const err = errors.command_is_conflict(existsParsedCommand.command, existsParsedCommand.clz.name, existsParsedCommand.location, parsedCommand.clz.name, parsedCommand.location);
-      if (!commandMeta.overrideCommand && !isInheritFrom(parsedCommand.clz, existsParsedCommand.clz)) {
-        throw err;
-      }
-
-      debug(err.message);
-    }
-
-    // handle middlewares
-    // Default orders:
-    //
-    // In class inheritance:
-    //              command1  <-extend-  command2
-    // trigger --> middleware1   -->   middleware2 --> middleware3  --> run
-    //
-    // ------------
-    //
-    // In run method:
-    //                      command2                               command1
-    // trigger --> middleware2 --> middleware3 --> run --> middleware1 --> super.run
-
-    // merge command middlewares with inherit command
-    const middlewareMeta = Reflect.getOwnMetadata(MetadataEnum.MIDDLEWARE, clz) as (MiddlewareMeta | undefined);
-    const commandMiddlewareConfigList = middlewareMeta?.configList || [];
-    if (inheritCommand && middlewareMeta?.inheritMetadata !== false) {
-      parsedCommand.addMiddlewares('command', { middleware: inheritCommand.commandMiddlewares });
-    }
-    commandMiddlewareConfigList.forEach(config => parsedCommand.addMiddlewares('command', config));
-
-    // add run middlewares, no need to merge with inherit command
-    const executionMiddlewareConfig = Reflect.getOwnMetadata(MetadataEnum.RUN_MIDDLEWARE, clz) as (MiddlewareMeta | undefined);
-    const executionMiddlewareConfigList = executionMiddlewareConfig?.configList || [];
-    executionMiddlewareConfigList.forEach(config => parsedCommand.addMiddlewares('execution', config));
-
-    // cache the instance
-    this.commands.set(parsedCommandInfo.uid, parsedCommand);
-    this.parsedCommandMap.set(clz, parsedCommand);
-    return parsedCommand;
-  }
-
-  private build() {
-    this.commands = new Map();
-    this.parsedCommandMap = new Map();
-    const parsedCommands = this.commandList
-      .map(clz => this.initParsedCommand(clz))
-      .filter(c => !!c) as ParsedCommand[];
-
-    // handle parent and childs
-    parsedCommands
-      .sort((a, b) => a.depth - b.depth)
-      .forEach(parsedCommand => {
-        let parent: ParsedCommand | undefined;
-        parsedCommand.cmds.forEach(cmd => {
-          // fullCmd is the key of this.commands
-          const fullCmd = parent ? parent.cmds.concat(cmd).join(' ') : cmd;
-
-          let cacheParsedCommand = this.commands.get(fullCmd);
-          if (!cacheParsedCommand) {
-            // create empty node
-            debug('Create empty command for \'%s\'', fullCmd);
-            cacheParsedCommand = new ParsedCommand(EmptyCommand, {
-              commandConfig: {},
-              parsedCommandInfo: parseCommand(fullCmd, this.binName),
-            });
-            this.commands.set(fullCmd, cacheParsedCommand);
-          }
-
-          if (!parent) {
-            this.root = parent = cacheParsedCommand;
-            return;
-          }
-
-          cacheParsedCommand.parent = parent;
-          parent.childs.push(cacheParsedCommand);
-          parent = cacheParsedCommand;
-        });
-      });
-  }
-
-  get(clz: typeof Command) {
-    return this.parsedCommandMap.get(clz);
-  }
-}
-
 @Injectable({ scope: ScopeEnum.SINGLETON })
 export class ParsedCommands {
   @Inject()
@@ -324,13 +42,15 @@ export class ParsedCommands {
   get tree(): ParsedCommandTree {
     if (!this[TREE_SYMBOL]) {
       const commandList = this.container.getInjectableByTag(MetadataEnum.COMMAND);
-      this[TREE_SYMBOL] = new ParsedCommandTree(this.binInfo.binName, commandList);
+      const tree = this.container.get(ParsedCommandTree);
+      tree.build(commandList);
+      this[TREE_SYMBOL] = tree;
     }
     return this[TREE_SYMBOL];
   }
 
   get root() {
-    return this.tree.root;
+    return this.tree.root!;
   }
 
   get commands() {
